分片插入数据

+++ sharding-jdbc-base/src/main/java/com/engrz/demo/sharding/jdbc/algorithm/PreciseModuloShardingDatabaseAlgorithm.java
@@ -5,10 +5,10 @@ import org.apache.shardingsphere.api.sharding.standard.PreciseShardingValue;
 
 import java.util.Collection;
 
-public class PreciseModuloShardingDatabaseAlgorithm implements PreciseShardingAlgorithm<Integer> {
+public class PreciseModuloShardingDatabaseAlgorithm implements PreciseShardingAlgorithm<Long> {
 
     @Override
-    public String doSharding(final Collection<String> databaseNames, final PreciseShardingValue<Integer> shardingValue) {
+    public String doSharding(final Collection<String> databaseNames, final PreciseShardingValue<Long> shardingValue) {
         for (String each : databaseNames) {
             if (each.endsWith(shardingValue.getValue() % 2 + "")) {
                 return each;
